Extract render dispatcher tracking to internal hook
