+++ src/react/hooks/internal/index.ts
@@ -1,4 +1,5 @@
 // These hooks are used internally and are not exported publicly by the library
 export { useDeepMemo } from "./useDeepMemo.js";
 export { useIsomorphicLayoutEffect } from "./useIsomorphicLayoutEffect.js";
+export { useRenderGuard } from "./useRenderGuard.js";
 export { __use } from "./__use.js";
